CLIP/requirements.txt
absl-py==2.1.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
comm==0.1.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.3
executing==0.8.3
fastjsonschema==2.18.1
filelock==3.13.1
fonttools==4.43.1
fqdn==1.5.1
fsspec==2024.2.0
ftfy==6.1.1
grad-cam==1.5.0
grpcio==1.48.2
h5py==3.11.0
huggingface-hub==0.23.3
idna==3.4
imageio==2.31.6
importlib-metadata==6.0.0
importlib-resources==6.1.0
ipykernel
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.1.0
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.9.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
MedPy==0.4.0
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.3
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.4
QtPy==2.4.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segment-anything==1.0
Send2Trash==1.8.2
setuptools==68.0.0
SimpleITK==2.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torchvision==0.18.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.41.2
triton==2.3.1
ttach==0.0.3
types-python-dateutil==2.8.19.14
typing_extensions==4.12.1
tzdata==2023.4
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.1.2
wheel==0.41.2
widgetsnbextension==4.0.9
zipp==3.11.0
